requirements.txt
fastapi==0.115.14
python-docx==1.2.0
python-multipart==0.0.20
uvicorn[standard]==0.35.0
